Updated database support

+++ FlightPatternDetection/Controllers/DebugController.cs
@@ -66,10 +66,10 @@ namespace FlightPatternDetection.Controllers
         [HttpGet("GetFlightExample")]
         public async Task<Flight> GetFlights(string flightid)
         {
-            if (await _context.Flights.AnyAsync(x => x.FlightId == int.Parse(flightid)))
+            if (await _context.Flights.AnyAsync(x => x.FlightId == long.Parse(flightid)))
             {
 
-                var testFlight = await _context.Flights.FirstAsync(x => x.FlightId == int.Parse(flightid));
+                var testFlight = await _context.Flights.FirstAsync(x => x.FlightId == long.Parse(flightid));
                 return testFlight;
             }
             else return null;
@@ -77,7 +77,7 @@ namespace FlightPatternDetection.Controllers
         }
 
         [HttpPost("CreateFlightExample")]
-        public async Task<Flight> CreateFlight(int flightId, string registration, string iCAO, string modeS, string callSign)
+        public async Task<Flight> CreateFlight(long flightId, string registration, string iCAO, string modeS, string callSign)
         {
             var newFlight = new Flight()
             {
@@ -110,7 +110,7 @@ namespace FlightPatternDetection.Controllers
         }
 
         [HttpPost("CreateHoldingPatternExample")]
-        public async Task<HoldingPattern> CreateHoldingPattern(int flightId, string fixpoint, int laps, Direction direction, double legDistance, double altitude)
+        public async Task<HoldingPattern> CreateHoldingPattern(long flightId, string fixpoint, int laps, Direction direction, double legDistance, double altitude)
         {
             var newHoldingPattern = new HoldingPattern()
             {
